Feat: fingertable
Implemented builing of fingertable from algortihms provided.
Builds all entires for each node and sets that nodes routing table with the build structure.

Created a helper class for each entry in the fingertable with appropriate fields. Interval is extracted from start and end fields.

Create a helper method for locating the next successor given a target (start) so that each successor in each entry points to the correct node